third_party/WebKit/LayoutTests/fast/dom/Window/Location/window-override-window-using-defineGetter.html

<html>
<head>
    <script src="../../../../resources/js-test.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    try {
        window.__defineGetter__("window", function() {
            return { location: "haxored" };
        });
    } catch (e) { }

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
</body>
</html>
